+++ src/components/layout/header/header-dropdown/header-dropdown-menu.tsx
@@ -1,5 +1,6 @@
 import { Link } from '@/components/ui/link';
 import { Menu } from 'antd';
+import React from 'react';
 import { headerLinkItems } from '../header-link-items';
 
 export const HeaderDropdownMenu = (
@@ -7,7 +8,7 @@ export const HeaderDropdownMenu = (
     {headerLinkItems
       .filter(({ isDropdown }) => isDropdown)
       .map(({ content, href, divider }, index) => (
-        <>
+        <React.Fragment key={`header-dropdown-menu-item-${index}`}>
           <Menu.Item
             key={`header-dropdown-menu-item-${index}`}
             className="hover:bg-blue-300"
@@ -15,7 +16,7 @@ export const HeaderDropdownMenu = (
             <Link href={href!}>{content}</Link>
           </Menu.Item>
           {divider && <Menu.Divider />}
-        </>
+        </React.Fragment>
       ))}
   </Menu>
 );