+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   - Component Type enabled: `'text' | 'switch' | 'check-group' | 'checkbox' | 'file-dropzone' | 'number' | 'radio' | 'select'`
   - sample field array page
 
+- Header content update
+
 ## 2021.11.22
 
 - Update of dependencies due Vercel indicated bugs